refactor: extract ProductivityScreen UI components and ViewModel hook
Debt Identified: The `frontend/app/(main)/productivity.tsx` was a "God Class" exceeding 800 lines, violating SRP by mixing complex state management, data fetching, filtering logic, and large nested widget trees within a single file.

Refactored Code:
- Decoupled business logic and state management into a dedicated hook `frontend/src/hooks/useProductivityViewModel.ts`.
- Decomposed the UI into modular sub-components within `frontend/src/components/productivity/`:
  - `ProductivityHeader.tsx`
  - `ProductivitySearch.tsx`
  - `ProductivityTabs.tsx`
  - `ProductivityTaskFilters.tsx`
  - `ProductivityEmptyState.tsx`
  - `ProductivityTodayPlan.tsx`
  - `ProductivityListItem.tsx`
- Refactored `ProductivityScreen` to consume the ViewModel and act purely as an orchestrator for the sub-components.

Structural Note: This strictly adheres to the Clean Architecture standard by abstracting application logic away from the UI layer. Extracting the `renderItem` mapping and list headers/emptys states stabilizes the `FlatList` to prevent unnecessary re-renders. The component is now significantly more readable, modular, and maintainable.
